@@ -12,14 +12,14 @@
   },
   "scripts": {
     "build": "tsc",
-    "clean": "node script/clean.js",
+    "clean": "node script/clean.mjs",
     "fix": "npm run lint:fix && npm run format:fix",
     "format": "prettier --check .",
     "format:fix": "prettier --write .",
     "lint": "eslint .",
     "lint:fix": "eslint --fix .",
     "snapshot:update": "npm run test -- -u",
-    "test": "jest --config test/jest.config.js",
+    "test": "jest --config jest.config.mjs",
     "all": "npm run clean && npm run fix && npm run build && npm run test"
   },
   "main": "dist/index.js",
@@ -28,30 +28,22 @@
     "test": "test"
   },
   "dependencies": {
-    "@datastructures-js/priority-queue": "5.4.0",
-    "winston": "^3.10.0"
+    "@datastructures-js/priority-queue": "6.3.1"
   },
   "devDependencies": {
-    "@tsconfig/recommended": "^1.0.1",
-    "@types/jest": "^27.5.1",
-    "@types/node": "^17.0.33",
-    "@typescript-eslint/eslint-plugin": "^7.5.0",
-    "@typescript-eslint/parser": "^7.5.0",
-    "colors": "^1.4.0",
-    "dotenv": "^16.0.1",
+    "@eslint/js": "^9.17.0",
+    "@tsconfig/recommended": "^1.0.8",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.10.2",
     "editorconfig": "^2.0.0",
-    "eslint": "^8.57.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-airbnb-typescript": "^18.0.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-jest": "^27.9.0",
-    "eslint-plugin-unused-imports": "^3.1.0",
-    "fs-extra": "^11.1.1",
-    "jest": "^29.6.1",
-    "nock": "^13.3.1",
-    "prettier": "^3.0.0",
-    "ts-jest": "^29.1.1",
-    "ts-node": "^10.7.0",
-    "typescript": "^5.4.3"
+    "eslint": "^9.17.0",
+    "eslint-plugin-unused-imports": "^4.1.4",
+    "fs-extra": "^11.2.0",
+    "jest": "^29.7.0",
+    "prettier": "^3.4.2",
+    "ts-jest": "^29.2.5",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.18.2"
   }
 }
